Authentication added to all the todos routes

+++ Project03/TodoApp/routers/todos.py
@@ -5,6 +5,7 @@ from typing import Annotated
 from sqlalchemy.orm import Session
 from starlette import status
 from pydantic import BaseModel, Field
+from .auth import get_current_user
 
 router = APIRouter(
     prefix="",
@@ -18,7 +19,8 @@ def get_db() :
         db.close()
 
 db_dependency = Annotated[Session, Depends(get_db)]
-      
+user_dependency = Annotated[dict, Depends(get_current_user)]
+    
 class TodoRequest(BaseModel) :
     title : str = Field(min_length = 3)
     description : str = Field(min_length = 3, max_length = 100)
@@ -28,31 +30,64 @@ class TodoRequest(BaseModel) :
     
 # -------- Read all todos ----------
 @router.get("/todos", status_code = status.HTTP_200_OK)
-async def read_all_todos(db: db_dependency) :
-    return db.query(Todos).all()
+async def read_all_todos(user : user_dependency, db: db_dependency) :
+    
+    if user is None :
+        raise HTTPException(
+            status_code = 401, 
+            detail = 'Authentication Failed'
+        )
+        
+    return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
+
+# ---------- Create new Todo -----------
+@router.post("/todos/todo", status_code = status.HTTP_201_CREATED)
+async def create_todo(user : user_dependency,
+                      db : db_dependency, 
+                      todo_request : TodoRequest) :
+    if user is None :
+        raise HTTPException(
+            status_code = 401, 
+            detail = 'Authentication Failed'
+        )
+        
+    todo_model = Todos(**todo_request.model_dump(), owner_id = user.get('id'))
+    
+    db.add(todo_model)
+    db.commit()
+    
+    return todo_model
 
 # -------- Read single todo ----------
 @router.get("/todos/{todo_id}", status_code = status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id : int = Path(gt=0)) :
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user : user_dependency, db: db_dependency, todo_id : int = Path(gt=0)) :
+    
+    if user is None :
+        raise HTTPException(
+            status_code = 401, 
+            detail = 'Authentication Failed'
+        )
+        
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
+    
     if todo_model is not None :
         return todo_model
     raise HTTPException(status_code=404, detail='Todo not found.')
 
-# ---------- Create new Todo -----------
-@router.post("/todos/todo", status_code = status.HTTP_201_CREATED)
-async def create_todo(db : db_dependency, todo_request : TodoRequest) :
-    todo_model = Todos(**todo_request.model_dump())
-    db.add(todo_model)
-    db.commit()
-    return todo_model
-
 # --------- Update todo ------------
-@router.put("/todos/{to_id}", status_code = status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency,
+@router.put("/todos/{todo_id}", status_code = status.HTTP_204_NO_CONTENT)
+async def update_todo(user : user_dependency,
+                      db: db_dependency,
                       todo_request : TodoRequest,
                       todo_id : int = Path(gt = 0)) :
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    
+    if user is None :
+        raise HTTPException(
+            status_code = 401, 
+            detail = 'Authentication Failed'
+        )
+        
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None :
         raise HTTPException(status_code=404, detail = 'Todo not found')
     
@@ -66,11 +101,22 @@ async def update_todo(db: db_dependency,
     
 # ------ Delete Todo ---------
 @router.delete("/todos/{todo_id}", status_code = status.HTTP_204_NO_CONTENT)
-async def delete_todo(db : db_dependency, todo_id : int = Path(gt=0)) :
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(
+        user : user_dependency, 
+        db : db_dependency, 
+        todo_id : int = Path(gt=0)) :
+    
+    if user is None :
+        raise HTTPException(
+            status_code = 401, 
+            detail = 'Authentication Failed'
+        )
+        
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
+    
     if todo_model is None :
         raise HTTPException(status_code = 404, detail = 'Todo not found')
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
     
     db.commit()
